Make IngredientCard action icons visible with explicit colors

+++ src/components/Dashboard_Manager/Storage/graphql/inventory.gql.js
@@ -269,7 +269,7 @@ export const ADJUST_STOCK = gql`
     $restaurantId: ID!
     $warehouseId: ID!
     $ingredientId: ID!
-    $qty: Float!
+    $qty: Int!
     $reason: String
   ) {
     adjustStock(
